store/requirements.txt
psycopg2-binary
sqlalchemy
alembic
pytz
marshmallow_dataclass
pyzmq
python-dateutil
flask

black
pytest
